Changed reflex and scroll.
When a shape is picked up the middle block of the array is selected as a base point to draw the shape from.

When the shape is scrolled, the array is changed, the base point should be updated again to the middle of the array to keep the center of the shape on the middle of the array.

^ currently /kinda/ working, breaks for certain shapes.

Issues where reflecting a shape of one type reflected them all should be fixed.