Dragging works
At the expense of reflecting and scrolling the shape, dragging is working how I want it to.

You click on a block that is part of the shape.
You drag the shape around with the block you picked up.

Scrolling changes the shapes pattern, but we are using the shapes blocks to draw it so nothing happens.

Now need a way to rotate the matrix while retaining the block that is being dragged.